updated hamr2 code with decoder tests 2 and 3

+++ hamr2/decoder.h
@@ -1,6 +1,6 @@
 #ifndef DECODER_h
 #define DECODER_h 
 
-int read_decoder(int index);
+short int read_decoder(int index);
 
 #endif
